chore: move vo to subscription namespace
