+++ src/Domain/Purchase/Subscription/SubscriptionState.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Imdhemy\GooglePlay\ValueObjects;
+namespace Imdhemy\GooglePlay\Domain\Purchase\Subscription;
 
 enum SubscriptionState: string
 {
