+++ src/Domain/Purchase/Subscription/TestPurchase.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Imdhemy\GooglePlay\ValueObjects;
+namespace Imdhemy\GooglePlay\Domain\Purchase\Subscription;
 
 /**
  * @see https://developers.google.cn/android-publisher/api-ref/rest/v3/purchases.subscriptionsv2#testpurchase
