+++ tests/Domain/Purchase/Subscription/PausedStateContextTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Tests\ValueObjects;
+namespace Tests\Domain\Purchase\Subscription;
 
-use Imdhemy\GooglePlay\ValueObjects\PausedStateContext;
+use Imdhemy\GooglePlay\Domain\Purchase\Subscription\PausedStateContext;
 use Tests\TestCase;
 
 final class PausedStateContextTest extends TestCase
