+++ tests/Domain/Purchase/Subscription/SubscriptionPurchaseLineItemTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Tests\ValueObjects;
+namespace Tests\Domain\Purchase\Subscription;
 
-use Imdhemy\GooglePlay\ValueObjects\SubscriptionPurchaseLineItem;
+use Imdhemy\GooglePlay\Domain\Purchase\Subscription\SubscriptionPurchaseLineItem;
 use Tests\TestCase;
 
 final class SubscriptionPurchaseLineItemTest extends TestCase
