+++ tests/Domain/Purchase/Subscription/TestPurchaseTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Tests\ValueObjects;
+namespace Tests\Domain\Purchase\Subscription;
 
-use Imdhemy\GooglePlay\ValueObjects\TestPurchase;
+use Imdhemy\GooglePlay\Domain\Purchase\Subscription\TestPurchase;
 use Tests\TestCase;
 
 final class TestPurchaseTest extends TestCase
